+++ src/main/java/net/oxyopia/vice/config/features/worlds/CookingConfig.java
@@ -1,7 +1,7 @@
 package net.oxyopia.vice.config.features.worlds;
 
 import com.google.gson.annotations.Expose;
-import net.oxyopia.vice.config.HudManager;
+import net.oxyopia.vice.data.Position;
 
 import java.util.HashMap;
 
@@ -20,6 +20,6 @@ public class CookingConfig {
 	public HashMap<String, Integer> totalBurgersComplete = new HashMap<>();
 
 	@Expose
-	public HudManager.Position currentOrderPos = new HudManager.Position(0.5f, 0.05f);
+	public Position currentOrderPos = new Position(175f, 10f);
 
 }
