+++ src/main/java/net/oxyopia/vice/data/EditableHudElement.kt
@@ -1,7 +1,5 @@
 package net.oxyopia.vice.data
 
-import net.oxyopia.vice.config.HudManager.Position
-
 abstract class EditableHudElement(val displayName: String, val defaultState: Position) {
 	abstract fun setPosition(x: Float, y: Float, scale: Float = 1f, centered: Boolean = true)
